market/realtime-prices/requirements.txt
asyncio==3.4.3
kafka-python==2.0.2
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
